backend/requirements.txt
fastapi
uvicorn
numpy
scipy
pydantic
yfinance
google-generativeai
